appeal updates (#18626)
* appeal updates

* appeal updates

* doc updates
